move ChannelList
